chore: 라우팅 경로 상수화
