+++ src/components/Buttons/CreateBtn.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
 import { Plus } from 'lucide-react'
 import Link from 'next/link'
+import { ROUTES } from '@/constants/routes'
 
 export default function CreateBtn() {
   return (
-    <Link href="/beauty-talk/create">
+    <Link href={ROUTES.beautyTalk.create}>
       <div>
         {/* 기본 버튼 */}
         <button className="flex items-center gap-2 px-4 py-2">
